dom/workers/test/test_WorkerDebugger.postMessage.xul

<?xml version="1.0"?>
<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<window title="Test for WorkerDebugger.postMessage"
        xmlns="http://www.mozilla.org/keymaster/gatekeeper/there.is.only.xul"
        onload="test();">

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"/>
  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/EventUtils.js"/>
  <script type="application/javascript" src="dom_worker_helper.js"/>

  <script type="application/javascript">
  <![CDATA[

    const WORKER_URL = "WorkerDebugger.postMessage_worker.js";
    const DEBUGGER_URL = BASE_URL + "WorkerDebugger.postMessage_debugger.js";

    function test() {
      Task.spawn(function* () {
        SimpleTest.waitForExplicitFinish();

        let promise = waitForRegister();
        let worker = new Worker(WORKER_URL);
        let dbg = yield promise;
        dbg.initialize(DEBUGGER_URL);

        promise = waitForDebuggerMessage(dbg);
        dbg.postMessage("ping");
        let message = yield promise;
        is(message, "pong", "pong message should be received");

        SimpleTest.finish();
      });
    }

  ]]>
  </script>

  <body xmlns="http://www.w3.org/1999/xhtml">
    <p id="display"></p>
    <div id="content" style="display:none;"></div>
    <pre id="test"></pre>
  </body>
  <label id="test-result"/>
</window>
